# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inventory/items:
    get:
      operationId: listInventoryItems
      tags: [inventory]
      summary: List the signed-in user's household items
      responses:
        "200":
          description: Household inventory
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/InventoryItem" } }
        "401": { description: Sign-in required, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Add a household item
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InventoryItemInput" }
      responses:
        "201":
          description: Item created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InventoryItem" }
        "400": { description: Invalid item, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "401": { description: Sign-in required, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
  /inventory/items/{itemId}:
    parameters:
      - name: itemId
        in: path
        required: true
        schema: { type: string, minLength: 1 }
    get:
      operationId: getInventoryItem
      tags: [inventory]
      summary: Get a household item with its records
      responses:
        "200":
          description: Household item
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InventoryItemDetail" }
        "401": { description: Sign-in required, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "404": { description: Item not found, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
    patch:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update a household item
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InventoryItemUpdate" }
      responses:
        "200":
          description: Item updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InventoryItem" }
        "400": { description: Invalid item, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "401": { description: Sign-in required, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "404": { description: Item not found, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
    delete:
      operationId: deleteInventoryItem
      tags: [inventory]
      summary: Delete a household item and its records
      responses:
        "204": { description: Item deleted }
        "401": { description: Sign-in required, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "404": { description: Item not found, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
  /inventory/items/{itemId}/diagnoses:
    parameters:
      - name: itemId
        in: path
        required: true
        schema: { type: string, minLength: 1 }
    get:
      operationId: listInventoryDiagnoses
      tags: [inventory]
      summary: List diagnosis history for an item
      responses:
        "200":
          description: Diagnosis history
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/DiagnosisResult" } }
        "401": { description: Sign-in required, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "404": { description: Item not found, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
  /inventory/items/{itemId}/documents:
    parameters:
      - name: itemId
        in: path
        required: true
        schema: { type: string, minLength: 1 }
    get:
      operationId: listInventoryDocuments
      tags: [inventory]
      summary: List private documents attached to an item
      responses:
        "200":
          description: Item documents
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/InventoryDocument" } }
        "401": { description: Sign-in required, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "404": { description: Item not found, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
    post:
      operationId: createInventoryDocument
      tags: [inventory]
      summary: Save metadata for an uploaded private document
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InventoryDocumentInput" }
      responses:
        "201":
          description: Document saved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InventoryDocument" }
        "400": { description: Invalid document, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "401": { description: Sign-in required, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "404": { description: Item not found, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
  /inventory/items/{itemId}/documents/{documentId}:
    parameters:
      - name: itemId
        in: path
        required: true
        schema: { type: string, minLength: 1 }
      - name: documentId
        in: path
        required: true
        schema: { type: string, minLength: 1 }
    delete:
      operationId: deleteInventoryDocument
      tags: [inventory]
      summary: Delete an item document
      responses:
        "204": { description: Document deleted }
        "401": { description: Sign-in required, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "404": { description: Document not found, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
  /inventory/items/{itemId}/reminders:
    parameters:
      - name: itemId
        in: path
        required: true
        schema: { type: string, minLength: 1 }
    get:
      operationId: listInventoryReminders
      tags: [inventory]
      summary: List maintenance reminders for an item
      responses:
        "200":
          description: Maintenance reminders
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/InventoryReminder" } }
        "401": { description: Sign-in required, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "404": { description: Item not found, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
    post:
      operationId: createInventoryReminder
      tags: [inventory]
      summary: Add an item maintenance reminder
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InventoryReminderInput" }
      responses:
        "201":
          description: Reminder created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InventoryReminder" }
        "400": { description: Invalid reminder, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "401": { description: Sign-in required, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "404": { description: Item not found, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
  /inventory/items/{itemId}/reminders/{reminderId}:
    parameters:
      - name: itemId
        in: path
        required: true
        schema: { type: string, minLength: 1 }
      - name: reminderId
        in: path
        required: true
        schema: { type: string, minLength: 1 }
    patch:
      operationId: updateInventoryReminder
      tags: [inventory]
      summary: Update or complete an item maintenance reminder
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InventoryReminderUpdate" }
      responses:
        "200":
          description: Reminder updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InventoryReminder" }
        "400": { description: Invalid reminder, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "401": { description: Sign-in required, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "404": { description: Reminder not found, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
...
  /inventory/items/{itemId}/repairs/{repairId}:
    parameters:
      - name: itemId
        in: path
        required: true
        schema: { type: string, minLength: 1 }
      - name: repairId
        in: path
        required: true
        schema: { type: string, minLength: 1 }
    patch:
      operationId: updateInventoryRepair
      tags: [inventory]
      summary: Update a before and after repair record
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InventoryRepairRecordUpdate" }
      responses:
        "200":
          description: Repair record updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InventoryRepairRecord" }
        "400": { description: Invalid repair record, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "401": { description: Sign-in required, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "404": { description: Repair record not found, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
...
components:
  schemas:
    InventoryItem:
      type: object
      required: [id, userId, category, name, manufacturer, model, serialNumber, purchaseDate, warrantyStartDate, warrantyEndDate, notes, createdAt, updatedAt]
      properties:
        id: { type: string }
        userId: { type: string }
        category:
          type: string
          enum: [kitchen_appliances, heating_equipment, tvs_electronics, plumbing_fixtures, garden_equipment]
        name: { type: string }
        manufacturer: { type: ["string", "null"] }
        model: { type: ["string", "null"] }
        serialNumber: { type: ["string", "null"] }
        purchaseDate: { type: ["string", "null"], format: date }
        warrantyStartDate: { type: ["string", "null"], format: date }
        warrantyEndDate: { type: ["string", "null"], format: date }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
...
    InventoryItemUpdate:
      allOf:
        - $ref: "#/components/schemas/InventoryItemInput"
      required: []
    InventoryItemDetail:
      type: object
      required: [item, documents, reminders, repairs]
      properties:
        item: { $ref: "#/components/schemas/InventoryItem" }
        documents: { type: array, items: { $ref: "#/components/schemas/InventoryDocument" } }
        reminders: { type: array, items: { $ref: "#/components/schemas/InventoryReminder" } }
        repairs: { type: array, items: { $ref: "#/components/schemas/InventoryRepairRecord" } }
...
    InventoryRepairRecordUpdate:
      allOf:
        - $ref: "#/components/schemas/InventoryRepairRecordInput"
      required: []
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error